requirements.txt
streamlit==1.38.0 
langchain-community==0.2.16 
langgraph==0.2.14 
fpdf==1.7.2 
python-dotenv==1.0.1
langchain-groq==0.3.4